aprs-is login:
aprs-is servers answers on login with i.e.
  # logresp DL9SAU-T1 unverified, server T2EISBERG

This answer was not intuitive and I checked the word " verified" for successful login.
In other aprs-is client software, I read they parsed "logresp" as answer from aprs-is servers.

=> My change

+           if (s.indexOf(" logresp") == -1) { aprsis_status = "Error: Login denied: " + s; aprsis_status.trim(); goto on_err; }
+           if (s.indexOf(" verified") == -1) { aprsis_status = "Notice: server responsed not verified: " + s; aprsis_status.trim();

An other response of a server is:
  user DL9SAU-TEST pass -1 scripttest 0.1
  # Invalid username format
We'll send Login denied here, with that message-
[Reason for deny is: len(SSID) is max 2 by definition of the aprs-is-authors.]

If password is wrong, we stay connected.
Notice message in aprsis_staus shows in the web-interface what went wrong.

Signed-off-by: Thomas Osterried <dl9sau@darc.de>